chatbot-web/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
openai==1.6.1
httpx>=0.24.1